fix: stack unwinding on linux (for real this time)
Ok, so the problem was not fixed in the previous commit
because it only worked if we unwinded first in the os thead's
stack. It seems that the first time you unwind the stack,
a lot of stack is used and a lot of worked is cached. This
meant that when we unwinded first on the os thread, then
it wouldn't stackoverflow in the coroutine. The solution was
to increase the stacksize.